requirements.txt
discord.py==2.0.1
python-dotenv==0.19.2
openai==0.27.4
numpy==1.21.2
pinecone-client==3.0.0